common/bitutil: optimize ORBytes
